Rename: gamePlayPageIcons 폴더명 띄어쓰기 제거
