+++ src/components/icons/gamePlayPageIcons/MinusIcon.tsx
