+++ src/components/icons/gamePlayPageIcons/MultiPlayerIcon.tsx
