+++ src/components/icons/gamePlayPageIcons/PlusIcon.tsx
