+++ src/components/icons/gamePlayPageIcons/SinglePlayerIcon.tsx
