+++ src/components/icons/gamePlayPageIcons/UserMinusIcon.tsx
